Added detail flag to test config
Added ['import']['detail'] flag to show individual track changes
so that unicode item changes can continue to be tested by previous
test. Adjusted assertions so that the changed fields are searched
for, rather than string including '->' as now the item length is
printed also.

Currently the output of the item changes are printed to a logger
for convenience to assist comparison with the old functionality
however this will be removed before the final merge.

ToDo:

 - Write tests for checking the new unicode characters printed
   by the changed UI.