Declared template for printing as Unicode
In commands.py, added a `u` prefix to string so it is treated as
Unicode for Python 2.7 compliance. Also removed test for Unicode
not equals character from ShowChangeTest (as a separate Unicode
test is required for these new characters if they are to be a
part of the UI). Only failing tests now (when run locally) are
MBSubmit raising a UnicodeDecode error when checking console
output against a string.