Moving track printing into agnostic Change class
To allow similarity in gui output between TrackMatch & AlbumMatch
diffs, we move the printing functions into the ChangeRepresentation
class, and extend it to either a track or album match as
appropriate.

The functions have not all been transferred or tested, extra ToDo
comments have been added to assist further work.